change data type for rating

+++ app/src/main/java/com/example/movievault/domain/model/Movie.kt
@@ -6,5 +6,5 @@ data class Movie(
     val posterPath: String?,
     val overview: String?,
     val releaseDate: String?,
-    val rating: Float
+    val rating: Double
 )
